refactor(reborn): split capability preview projection
