+++ crates/ironclaw_product_adapters/src/lib.rs
@@ -50,7 +50,9 @@ pub use inbound::{
     TrustedInboundContext, UserMessagePayload, parse_product_slash_command,
 };
 pub use outbound::{
-    AuthPromptView, CapabilityActivityStatusView, CapabilityActivityView,
+    AuthPromptView, CAPABILITY_DISPLAY_KIND_MAX_BYTES, CAPABILITY_DISPLAY_PREVIEW_MAX_BYTES,
+    CAPABILITY_DISPLAY_PREVIEW_MAX_LINES, CAPABILITY_DISPLAY_RESULT_REF_MAX_BYTES,
+    CAPABILITY_DISPLAY_SUMMARY_MAX_BYTES, CapabilityActivityStatusView, CapabilityActivityView,
     CapabilityActivityViewInput, CapabilityDisplayPreviewView, CapabilityDisplayPreviewViewInput,
     FinalReplyView, GatePromptView, ProductOutboundEnvelope, ProductOutboundPayload,
     ProductOutboundTarget, ProductProjectionItem, ProductProjectionState, ProductRenderOutcome,
