+++ crates/ironclaw_product_adapters/src/outbound.rs
@@ -19,11 +19,11 @@ const PROJECTION_TEXT_MAX_BYTES: usize = 128 * 1024;
 const CAPABILITY_ACTIVITY_ERROR_KIND_MAX_BYTES: usize = 64;
 const CAPABILITY_ACTIVITY_ERROR_KIND_SEGMENT_MAX_BYTES: usize = 24;
 const CAPABILITY_ACTIVITY_UNCLASSIFIED_ERROR_KIND: &str = "Unclassified";
-const CAPABILITY_DISPLAY_SUMMARY_MAX_BYTES: usize = 2 * 1024;
-const CAPABILITY_DISPLAY_PREVIEW_MAX_BYTES: usize = 16 * 1024;
-const CAPABILITY_DISPLAY_PREVIEW_MAX_LINES: usize = 120;
-const CAPABILITY_DISPLAY_KIND_MAX_BYTES: usize = 32;
-const CAPABILITY_DISPLAY_RESULT_REF_MAX_BYTES: usize = 256;
+pub const CAPABILITY_DISPLAY_SUMMARY_MAX_BYTES: usize = 2 * 1024;
+pub const CAPABILITY_DISPLAY_PREVIEW_MAX_BYTES: usize = 16 * 1024;
+pub const CAPABILITY_DISPLAY_PREVIEW_MAX_LINES: usize = 120;
+pub const CAPABILITY_DISPLAY_KIND_MAX_BYTES: usize = 32;
+pub const CAPABILITY_DISPLAY_RESULT_REF_MAX_BYTES: usize = 256;
 
 fn invalid(kind: &'static str, reason: impl Into<String>) -> ProductAdapterError {
     ProductAdapterError::InvalidIdentifier {
