+++ crates/ironclaw_reborn_composition/src/projection/tests.rs
@@ -22,6 +22,7 @@ use ironclaw_turns::{
 };
 
 mod cursor_validation;
+mod display_preview;
 mod runtime_stream;
 mod turn_stream;
